PS) DynamicProgramming 11727

+++ DynamicProgramming/Main2.java
@@ -5,17 +5,17 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 
 public class Main2 {
-  static int[] arr;
+  static long[] arr;
   public static void main(String[] args) throws IOException {
     BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(System.in));
     int N = Integer.parseInt(bufferedReader.readLine());
 
-    arr = new int[N];
+    arr = new long[N];
 
     System.out.println(prinaryNumber(N-1));
   }
 
-  public static int prinaryNumber(int num){
+  public static long prinaryNumber(int num){
     if(num == 0 || num == 1){
       return arr[num] = 1;
     }
